Put all of maybe effect example in one file
